Limiting the Search Space in Optimal Quantum Circuit Mapping (#6)
 - ⚡ actually enables -march=native -mtune=native for source builds via pip
 - ⚡ setup.py now runs on as many threads as available
 - 🏗️ switch to PEP 517 `pyproject.toml`
 - 💚 switch cibuildwheel configuration to `pyproject.toml`
 - 🚀 added Python 3.10 wheels
 - ♻️ refactor the Python bindings to make them easier to use
 - ⬆️ updated all external submodules to their most recent version
 - 📝 updated documentation

+++ src/utils.cpp
@@ -140,5 +140,5 @@ unsigned long idx(unsigned int k, unsigned short i, unsigned short j, const std:
         counti++;
     }
 
-    return k * nj * iValues.size() + counti * nj + j;
+    return k * static_cast<std::size_t>(nj) * iValues.size() + counti * static_cast<std::size_t>(nj) + j;
 }
